requirements.txt
# AI Integration
google-genai>=1.0.0

# MCP Integration
mcp>=1.0.0

# Terminal UI
textual>=0.50.0

# Configuration Management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0